[merge] 머지 컨플릭트 해결

+++ src/main/java/camp/woowak/lab/store/service/response/StoreDisplayResponse.java
@@ -53,7 +53,7 @@ public record StoreDisplayResponse(
 
 		Long menuId,
 		String menuName,
-		Integer menuPrice
+		Long menuPrice
 	) {
 
 		public static MenuDisplayResponse of(final Menu menu) {
